@@ -5,37 +5,36 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.5.3</version>
+        <version>3.2.11</version>
         <relativePath/>
     </parent>
     <groupId>top.slcp</groupId>
     <artifactId>myblog</artifactId>
-    <version>2.0.0.SNAPSHOT</version>
+    <version>3.0.0.SNAPSHOT</version>
     <packaging>jar</packaging>
     <name>slcp</name>
-    <description>Blog project for Spring Boot</description>
+    <description>Blog project for Spring Boot 3 with Spring AI</description>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
-        <qiniu.version>7.2.6</qiniu.version>
-        <httpclient.version>4.5.12</httpclient.version>
-        <aliyun.version>4.5.3</aliyun.version>
+        <java.version>17</java.version>
+        <spring-ai.version>0.8.1</spring-ai.version>
+        <qiniu.version>7.15.1</qiniu.version>
+        <httpclient.version>5.3</httpclient.version>
+        <aliyun.version>4.6.4</aliyun.version>
 
-        <swagger.version>2.10.5</swagger.version>
-        <swagger.models.version>1.6.2</swagger.models.version>
-        <knife4j.version>2.0.8</knife4j.version>
-        <fastjson.version>1.2.83</fastjson.version>
-        <commonmark.version>0.10.0</commonmark.version>
-        <alipay.version>3.0.0</alipay.version>
-        <hutool.version>5.8.1</hutool.version>
-        <redis.version>3.6.3</redis.version>
-        <redisson.version>3.16.1</redisson.version>
-        <mybatisplus.version>3.4.3.2</mybatisplus.version>
-        <druid.version>1.2.6</druid.version>
+        <springdoc.version>2.3.0</springdoc.version>
+        <knife4j.version>4.4.0</knife4j.version>
+        <fastjson.version>2.0.43</fastjson.version>
+        <commonmark.version>0.17.0</commonmark.version>
+        <alipay.version>4.38.10.ALL</alipay.version>
+        <hutool.version>5.8.24</hutool.version>
+        <jedis.version>5.1.0</jedis.version>
+        <redisson.version>3.26.0</redisson.version>
+        <mybatisplus.version>3.5.7</mybatisplus.version>
+        <druid.version>1.2.20</druid.version>
         <algolia.version>3.16.5</algolia.version>
-        <openai.version>0.9.0</openai.version>
     </properties>
 
     <dependencies>
