@@ -1,14 +1,7 @@
 package com.slcp.devops;
 
-import com.plexpt.chatgpt.ChatGPT;
-import com.plexpt.chatgpt.entity.chat.ChatCompletion;
-import com.plexpt.chatgpt.entity.chat.ChatCompletionResponse;
-import com.plexpt.chatgpt.entity.chat.Message;
-import com.plexpt.chatgpt.util.Proxys;
+import com.slcp.devops.service.AiChatService;
 import com.slcp.devops.service.QqUserService;
-
-import com.slcp.devops.utils.OpenAiUtils;
-import com.theokanning.openai.completion.CompletionChoice;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -16,13 +9,10 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.MimeMessageHelper;
 
-import javax.mail.MessagingException;
-import javax.mail.internet.MimeMessage;
-import java.net.Proxy;
+import jakarta.mail.MessagingException;
+import jakarta.mail.internet.MimeMessage;
 import java.text.SimpleDateFormat;
-import java.util.Arrays;
 import java.util.Date;
-import java.util.List;
 
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@@ -32,89 +22,42 @@ class DevopsApplicationTests {
     @Autowired
     private JavaMailSender javaMailSender;
 
-    @Value("${openai.token}")
-    private String OPENAPI_TOKEN;
+    @Autowired(required = false)
+    private AiChatService aiChatService;
 
     /**
-     * 测试问答
+     * 测试问答 - 使用 Spring AI
      */
     @Test
     public void testQA(){
-        List<CompletionChoice> questionAnswer = OpenAiUtils.getQuestionAnswer("长沙有哪些好玩的地方？");
-        for (CompletionChoice completionChoice : questionAnswer) {
-            System.out.println(completionChoice.getText());
+        if (aiChatService != null) {
+            String response = aiChatService.chat("长沙有哪些好玩的地方？");
+            System.out.println(response);
+        } else {
+            System.out.println("AiChatService not configured, skipping test");
         }
     }
 
     /**
-     * 测试面试题生成
+     * 测试面试题生成 - 使用 Spring AI
      */
     @Test
     public void testInterview(){
-        /*//国内访问需要做代理，国外服务器不需要
-        Proxy proxy = new Proxy(Proxy.Type.HTTP, new InetSocketAddress("127.0.0.1", 7890));
-        HttpLoggingInterceptor httpLoggingInterceptor = new HttpLoggingInterceptor(new OpenAILogger());
-        //！！！！千万别再生产或者测试环境打开BODY级别日志！！！！
-        //！！！生产或者测试环境建议设置为这三种级别：NONE,BASIC,HEADERS,！！！
-        httpLoggingInterceptor.setLevel(HttpLoggingInterceptor.Level.HEADERS);
-        OkHttpClient okHttpClient = new OkHttpClient
-                .Builder()
-                .proxy(proxy)//自定义代理
-                .addInterceptor(httpLoggingInterceptor)//自定义日志输出
-                .addInterceptor(new OpenAiResponseInterceptor())//自定义返回值拦截
-                .connectTimeout(15, TimeUnit.SECONDS)//自定义超时时间
-                .writeTimeout(1200, TimeUnit.SECONDS)//自定义超时时间
-                .readTimeout(1200, TimeUnit.SECONDS)//自定义超时时间
-                .build();
-        //构建客户端
-        OpenAiClient openAiClient = OpenAiClient.builder()
-                .apiKey(OPENAPI_TOKEN)
-                .okHttpClient(okHttpClient)
-                .build();
-        //简单模型
-        //CompletionResponse completions = //openAiClient.completions("我想申请转专业，从计算机专业转到会计学专业，帮我完成一份两百字左右的申请书");
-        //最新GPT-3.5-Turbo模型
-        Message message = Message.builder().role(Message.Role.USER).content("java怎么学呀").build();
-        ChatCompletion chatCompletion = ChatCompletion.builder().messages(Arrays.asList(message)).build();
-        ChatCompletionResponse chatCompletionResponse = openAiClient.chatCompletion(chatCompletion);
-        StringBuilder st = new StringBuilder();
-        chatCompletionResponse.getChoices().forEach(e -> {
-            st.append(e.getMessage().getContent());
-        });
-        System.out.println(st);*/
-        //国内需要代理 国外不需要
-        Proxy proxy = Proxys.http("127.0.0.1", 7890);
-        ChatGPT chatGPT = ChatGPT.builder()
-                .apiKey(OPENAPI_TOKEN)
-                .proxy(proxy)
-                .timeout(900)
-                .apiHost("https://api.openai.com/") //反向代理地址
-                .build()
-                .init();
-
-        Message system = Message.ofSystem("你现在是一个博主，专门回复留言");
-        Message message = Message.of("java怎么学呀");
-
-        ChatCompletion chatCompletion = ChatCompletion.builder()
-                .model(ChatCompletion.Model.GPT_3_5_TURBO.getName())
-                .messages(Arrays.asList(system, message))
-                .maxTokens(3000)
-                .temperature(0.9)
-                .build();
-        ChatCompletionResponse response = chatGPT.chatCompletion(chatCompletion);
-        Message res = response.getChoices().get(0).getMessage();
-        System.out.println(res);
+        if (aiChatService != null) {
+            String response = aiChatService.chat("java怎么学呀");
+            System.out.println(response);
+        } else {
+            System.out.println("AiChatService not configured, skipping test");
+        }
     }
 
     @Test
     public void test() {
-        List<CompletionChoice> questionAnswer = OpenAiUtils.getQuestionAnswer("使用SpringBoot框架进行Http请求的详细代码");
-        for (CompletionChoice completionChoice : questionAnswer) {
-            System.out.println(completionChoice.getText());
-        }
-        List<CompletionChoice> openAiApi = OpenAiUtils.getOpenAiApi("使用SpringBoot框架进行Http请求");
-        for (CompletionChoice completionChoice : openAiApi) {
-            System.out.println(completionChoice.getText());
+        if (aiChatService != null) {
+            String response = aiChatService.chat("使用SpringBoot框架进行Http请求的详细代码");
+            System.out.println(response);
+        } else {
+            System.out.println("AiChatService not configured, skipping test");
         }
     }
 
