@@ -1,6 +1,5 @@
 package com.slcp.devops;
 
-import cn.hutool.core.util.StrUtil;
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -9,9 +8,11 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 /**
  * @author Slcp
  */
-@SpringBootApplication
+@SpringBootApplication(exclude = {
+    org.springframework.cloud.function.context.config.ContextFunctionCatalogAutoConfiguration.class
+})
 @EnableScheduling
-@MapperScan(basePackages="com.slcp.devops.mapper")
+@MapperScan(basePackages = "com.slcp.devops.mapper")
 public class DevopsApplication {
 
     public static void main(String[] args) {
