@@ -1,8 +1,6 @@
 package com.slcp.devops.config;
 
-import com.slcp.devops.utils.OpenAiUtils;
 import lombok.Data;
-import org.springframework.beans.factory.InitializingBean;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 import java.util.List;
@@ -10,11 +8,13 @@ import java.util.List;
 /**
  * @author slcp
  * @date 2023-01-10 22:10:43
+ * @deprecated 已废弃，请使用新的 Spring AI 配置
+ * 该类仅保留用于向后兼容，将在未来版本中移除
  */
-
+@Deprecated
 @Data
 @ConfigurationProperties(prefix = "openai")
-public class OpenAiProperties implements InitializingBean {
+public class OpenAiProperties {
     /**
      * 秘钥
      */
@@ -23,14 +23,4 @@ public class OpenAiProperties implements InitializingBean {
      * 超时时间
      */
     Integer timeout;
-
-    /**
-     * 设置属性时同时设置给OpenAiUtils
-     * @throws Exception 异常
-     */
-    @Override
-    public void afterPropertiesSet() throws Exception {
-        OpenAiUtils.OPENAPI_TOKEN = token;
-        OpenAiUtils.TIMEOUT = timeout;
-    }
 }
