@@ -1,11 +1,11 @@
 package com.slcp.devops.controller;
 
 import com.alibaba.fastjson.JSONObject;
+import com.slcp.devops.entity.Music;
 import com.slcp.devops.service.AboutService;
 import com.slcp.devops.service.IMusicService;
-import com.slcp.devops.entity.Music;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -21,7 +21,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
  */
 @Controller
 @AllArgsConstructor
-@Api(value = "关于我接口查询", tags = "关于我展示接口查询")
+@Tag(name = "关于我展示接口查询", description = "关于我接口查询")
 public class AboutShowController {
 
     private final AboutService aboutService;
@@ -39,7 +39,7 @@ public class AboutShowController {
 
     @GetMapping("/about/sChart")
     @ResponseBody
-    @ApiOperation(httpMethod = "GET", value = "quantityStatistics", notes = "文章、标签、分类数量统计")
+    @Operation(summary = "quantityStatistics", description = "文章、标签、分类数量统计")
     public JSONObject quantityStatistics(){
         return aboutService.getQuantityStatistics();
     }
