@@ -6,15 +6,14 @@ import com.alibaba.fastjson.JSONObject;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.slcp.devops.api.Result;
 import com.slcp.devops.config.DoQueryCache;
-import com.slcp.devops.constant.DevOpsConstant;
 import com.slcp.devops.dto.MtWallsDTO;
 import com.slcp.devops.entity.*;
 import com.slcp.devops.service.IMomentTimeService;
 import com.slcp.devops.service.IMtFeedbacksService;
 import com.slcp.devops.service.IMtWallsService;
 import com.slcp.devops.utils.ImgUtil;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
@@ -33,7 +32,7 @@ import java.util.Map;
 @RequestMapping("momentTime")
 @Slf4j
 @CrossOrigin
-@Api(value = "一刻时光", tags = "一刻时光")
+@Tag(name = "一刻时光", description = "一刻时光")
 public class MomentTimeController {
     private final IMomentTimeService momentTimeService;
     private final IMtWallsService mtWallsService;
@@ -44,7 +43,7 @@ public class MomentTimeController {
      * @return 返回
      */
     @PostMapping("/signip")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_POST, value = "signip", notes = "获取IP")
+    @Operation(summary = "signip", description = "获取IP")
     public Result<JSONObject> signIp() {
         return Result.data(ImgUtil.getIp());
     }
@@ -55,7 +54,7 @@ public class MomentTimeController {
      * @return 返回
      */
     @PostMapping("/insertfeedback")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_POST, value = "insertfeedback", notes = "查询墙")
+    @Operation(summary = "insertfeedback", description = "查询墙")
     public Result<MtFeedbacks> saveFeedBack(@RequestBody MtFeedbacks feedbacks) {
         feedbacksService.saveOrUpdate(feedbacks);
         return Result.data(feedbacks);
@@ -67,7 +66,7 @@ public class MomentTimeController {
      * @return 返回
      */
     @PostMapping("/insertwall")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_POST, value = "insertwall", notes = "新建数据")
+    @Operation(summary = "insertwall", description = "新建数据")
     public Result<MtWalls> saveWall(@RequestBody MtWalls mtWalls) {
         return Result.data(momentTimeService.saveWall(mtWalls));
     }
@@ -79,7 +78,7 @@ public class MomentTimeController {
      * @return 返回
      */
     @GetMapping("/findwallpage")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_GET, value = "getTypeById", notes = "根据主键获取信息")
+    @Operation(summary = "getTypeById", description = "根据主键获取信息")
     public Result<IPage<MtWallsDTO>> findWallPage(@RequestParam Map<String, Object> queryParam, Search search) {
         IPage<MtWallsDTO> mtWallsIPage = SqlWhereWrapper.getPage(search);
         MtWalls mtWalls = SqlWhereWrapper.convertParamsMapToObject(queryParam, MtWalls.class);
@@ -92,7 +91,7 @@ public class MomentTimeController {
      * @return 返回
      */
     @PostMapping("/insertcomment")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_POST, value = "insertcomment", notes = "新建评论")
+    @Operation(summary = "insertcomment", description = "新建评论")
     public Result<MtComments> saveComment(@RequestBody MtComments mtComments) {
         momentTimeService.saveOrUpdate(mtComments);
         return Result.data(mtComments);
@@ -105,7 +104,7 @@ public class MomentTimeController {
      * @return 返回
      */
     @GetMapping("/findcommentpage")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_GET, value = "findcommentpage", notes = "获取评论")
+    @Operation(summary = "findcommentpage", description = "获取评论")
     @DoQueryCache(expireTime = 5)
     public Result<IPage<MtComments>> findComment(@RequestParam Serializable id, Search search) {
         IPage<MtComments> mtCommentsIPage = SqlWhereWrapper.getPage(search);
@@ -126,7 +125,7 @@ public class MomentTimeController {
      * @return 返回
      */
     @PostMapping("/profile")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_POST, value = "profile", notes = "图片上传")
+    @Operation(summary = "profile", description = "图片上传")
     public Result<IPage<MtComments>> profile(@RequestBody MultipartFile file) {
         if (ObjectUtil.isEmpty(file)) {
             return Result.fail("上传文件为空~");
