@@ -2,11 +2,10 @@ package com.slcp.devops.controller.admin;
 
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.slcp.devops.api.Result;
-import com.slcp.devops.constant.DevOpsConstant;
 import com.slcp.devops.dto.BlogQueryDTO;
 import com.slcp.devops.entity.*;
 import com.slcp.devops.service.IBlogService;
-import io.swagger.annotations.ApiOperation;
+import io.swagger.v3.oas.annotations.Operation;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
@@ -28,26 +27,26 @@ public class BlogController {
     private final IBlogService blogService;
 
     @GetMapping("/blog")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_GET, value = "blogList", notes = "获取博客列表")
+    @Operation(summary = "blogList", description = "获取博客列表")
     public Result<IPage<BlogQueryDTO>> blogList(@RequestParam Map<String, Object> queryParam, Search search) {
         IPage<BlogQueryDTO> listInfoByPage = SqlWhereWrapper.getPage(search);
         return Result.data(blogService.listBlogs(listInfoByPage, (String) queryParam.get("query")));
     }
 
     @GetMapping("/blog/{id}")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_GET, value = "getBlogById", notes = "根据主键获取信息")
+    @Operation(summary = "getBlogById", description = "根据主键获取信息")
     public Result<Blog> getBlogById(@PathVariable("id") Long id) {
         return Result.data(blogService.getById(id));
     }
 
     @GetMapping("/blog/{id}/{typeId}")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_GET, value = "deleteBlog", notes = "删除分类")
+    @Operation(summary = "deleteBlog", description = "删除分类")
     public Result<Blog> deleteBlog(@PathVariable("id") Long id, @PathVariable("typeId") Long typeId) {
         return Result.condition(blogService.lambdaUpdate().set(Blog::getTypeId, typeId).eq(Blog::getId, id).update());
     }
 
     @PostMapping("/blog")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_POST, value = "addBlog", notes = "添加&修改分类")
+    @Operation(summary = "addBlog", description = "添加&修改分类")
     public Result<Blog> addBlog(@RequestBody Blog blog) {
         return Result.condition(blogService.saveOrUpdate(blog));
     }
