@@ -4,26 +4,21 @@ import cn.hutool.core.util.ObjectUtil;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.slcp.devops.api.Result;
-import com.slcp.devops.constant.DevOpsConstant;
 import com.slcp.devops.entity.Picture;
 import com.slcp.devops.entity.Search;
 import com.slcp.devops.entity.SqlWhereWrapper;
-import com.slcp.devops.utils.ImgUtil;
-import com.slcp.devops.utils.QiniuUtils;
 import com.slcp.devops.service.IPictureService;
+import com.slcp.devops.utils.ImgUtil;
 import com.slcp.devops.utils.StringUtil;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
 import java.util.Map;
-import java.util.UUID;
 
 /**
  * @author: Slcp
@@ -34,33 +29,33 @@ import java.util.UUID;
 @RestController
 @RequestMapping("/sys")
 @AllArgsConstructor
-@Api(value = "图片管理接口", tags = "图片管理接口")
+@Tag(name = "图片管理接口", description = "图片管理接口")
 @Slf4j
 public class PictureController {
 
     private final IPictureService pictureService;
 
     @GetMapping("/picture/{id}")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_GET, value = "getPictureById", notes = "根据主键获取信息")
+    @Operation(summary = "getPictureById", description = "根据主键获取信息")
     public Result<Picture> getPictureById(@PathVariable("id") Long id) {
         return Result.data(pictureService.getById(id));
     }
 
     @DeleteMapping("/picture/{id}")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_DELETE, value = "deletePicture", notes = "删除友链")
+    @Operation(summary = "deletePicture", description = "删除友链")
     public Result<Picture> deletePicture(@PathVariable("id") Long id) {
         StringUtil.delPicture(pictureService.getById(id).getPictureAddress());
         return Result.condition(pictureService.removeById(id));
     }
 
     @PostMapping("/picture")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_POST, value = "addPicture", notes = "添加&修改友链")
+    @Operation(summary = "addPicture", description = "添加&修改友链")
     public Result<Picture> addPicture(@RequestBody Picture picture) {
         return Result.condition(pictureService.saveOrUpdate(picture));
     }
 
     @GetMapping("/picture")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_GET, value = "pictureList", notes = "获取友链")
+    @Operation(summary = "pictureList", description = "获取友链")
     public Result<IPage<Picture>> pictureList(@RequestParam Map<String, Object> queryParam, Search search) {
         IPage<Picture> listInfoByPage = SqlWhereWrapper.getPage(search);
         String pictureName = (String) queryParam.get("query");
