@@ -4,14 +4,13 @@ import cn.hutool.core.util.ObjectUtil;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.slcp.devops.api.Result;
-import com.slcp.devops.constant.DevOpsConstant;
-import com.slcp.devops.entity.Type;
 import com.slcp.devops.entity.Search;
 import com.slcp.devops.entity.SqlWhereWrapper;
+import com.slcp.devops.entity.Type;
 import com.slcp.devops.service.ITypeService;
 import com.slcp.devops.utils.StringUtil;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
@@ -28,25 +27,25 @@ import java.util.Map;
 @RestController
 @RequestMapping("/sys")
 @AllArgsConstructor
-@Api(value = "类型接口查询", tags = "类型接口查询")
+@Tag(name = "类型接口查询", description = "类型接口查询")
 public class TypeController {
 
     private final ITypeService typeService;
 
     @GetMapping("/type/{id}")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_GET, value = "getTypeById", notes = "根据主键获取信息")
+    @Operation(summary = "getTypeById", description = "根据主键获取信息")
     public Result<Type> getTypeById(@PathVariable("id") Long id) {
         return Result.data(typeService.getById(id));
     }
 
     @DeleteMapping("/type/{id}")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_DELETE, value = "deleteType", notes = "删除分类")
+    @Operation(summary = "deleteType", description = "删除分类")
     public Result<Type> deleteType(@PathVariable("id") Long id) {
         return Result.condition(typeService.removeById(id));
     }
 
     @PostMapping("/type")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_POST, value = "addType", notes = "添加&修改分类")
+    @Operation(summary = "addType", description = "添加&修改分类")
     public Result<Type> addType(@RequestBody Type type) {
         if (ObjectUtil.isEmpty(type.getId()) && ObjectUtil.isNotEmpty(typeService.lambdaQuery().eq(Type::getName, type.getName()).one())) {
             return Result.fail("分类名已存在哦~");
@@ -55,13 +54,13 @@ public class TypeController {
     }
 
     @GetMapping("/types")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_GET, value = "types", notes = "获取所有分类")
+    @Operation(summary = "types", description = "获取所有分类")
     public Result<List<Type>> types() {
         return Result.data(typeService.list());
     }
 
     @GetMapping("/type")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_GET, value = "typeList", notes = "获取分类列表")
+    @Operation(summary = "typeList", description = "获取分类列表")
     public Result<IPage<Type>> typeList(@RequestParam Map<String, Object> queryParam, Search search) {
         IPage<Type> listInfoByPage = SqlWhereWrapper.getPage(search);
         String name = (String) queryParam.get("query");
