@@ -3,13 +3,12 @@ package com.slcp.devops.controller.admin;
 import cn.hutool.core.util.ObjectUtil;
 import com.alibaba.fastjson.JSONObject;
 import com.slcp.devops.api.Result;
-import com.slcp.devops.constant.DevOpsConstant;
 import com.slcp.devops.entity.SysAdmin;
 import com.slcp.devops.service.IAdminService;
 import com.slcp.devops.utils.ImgUtil;
 import com.slcp.devops.utils.MD5Utils;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
@@ -24,13 +23,13 @@ import java.time.LocalDateTime;
 @RestController
 @RequestMapping(value = "/sys")
 @AllArgsConstructor
-@Api(value = "用户接口", tags = "用户接口")
+@Tag(name = "用户接口", description = "用户接口")
 public class UserController {
 
     private final IAdminService adminService;
 
     @PostMapping("/login")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_POST, value = "login", notes = "登录")
+    @Operation(summary = "login", description = "登录")
     public Result<SysAdmin> login(@RequestBody SysAdmin admin) {
         SysAdmin sysAdmin = adminService.lambdaQuery().eq(SysAdmin::getUsername, admin.getUsername()).eq(SysAdmin::getPassword, MD5Utils.code(admin.getPassword())).one();
         if (ObjectUtil.isEmpty(sysAdmin)) {
@@ -44,7 +43,7 @@ public class UserController {
     }
 
     @GetMapping("/user/map/{id}")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_GET, value = "map", notes = "获取位置")
+    @Operation(summary = "map", description = "获取位置")
     public void map(@PathVariable("id") Long id) {
         //获取当前地理位置
         JSONObject ipJson = ImgUtil.getIp();
@@ -52,7 +51,7 @@ public class UserController {
     }
 
     @PostMapping("/users")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_POST, value = "saveOrUpdateOfUser", notes = "用户保存修改")
+    @Operation(summary = "saveOrUpdateOfUser", description = "用户保存修改")
     public Result<SysAdmin> saveOrUpdateOfUser(@RequestBody SysAdmin admin) {
         boolean flag = ObjectUtil.isEmpty(admin.getId());
         SysAdmin one = adminService.lambdaQuery().eq(SysAdmin::getUsername, admin.getUsername()).one();
@@ -68,20 +67,20 @@ public class UserController {
     }
 
     @GetMapping("/users/{id}")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_GET, value = "getUserOfId", notes = "根据主键获取用户信息")
+    @Operation(summary = "getUserOfId", description = "根据主键获取用户信息")
     public Result<SysAdmin> getUserOfId(@PathVariable("id") Long id) {
         return Result.data(adminService.getById(id));
     }
 
     @DeleteMapping("/users/{id}")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_DELETE, value = "deleteUser", notes = "删除用户")
+    @Operation(summary = "deleteUser", description = "删除用户")
     public Result<SysAdmin> deleteUser(@PathVariable("id") Long id) {
         return Result.condition(adminService.removeById(id));
     }
 
 
     @GetMapping("/roles/{id}/{rid}")
-    @ApiOperation(httpMethod = DevOpsConstant.METHOD_GET, value = "updRoleIdOfUser", notes = "根据用户id修改角色")
+    @Operation(summary = "updRoleIdOfUser", description = "根据用户id修改角色")
     public Result<SysAdmin> updRoleIdOfUser(@PathVariable("id") Long id, @PathVariable("rid") Long rid) {
         return Result.condition(adminService.lambdaUpdate().eq(SysAdmin::getId, id).set(SysAdmin::getRid, rid).update());
     }
