@@ -5,7 +5,7 @@ import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 import lombok.Getter;
 import org.springframework.format.annotation.DateTimeFormat;
@@ -25,14 +25,14 @@ public class BaseEntity<T> implements Serializable {
      * 博客不走集群部署，因此采用自增即可，若想高可用，集群部署，那么采用雪花算法
      */
     @JsonSerialize(using = ToStringSerializer.class)
-    @ApiModelProperty(value = "主键id")
+    @Schema(description = "主键id")
     @TableId(value = "id")
     private Long id;
 
     /**
      * 创建时间
      */
-    @ApiModelProperty(value = "创建时间")
+    @Schema(description = "创建时间")
     @TableField(value = "create_time", fill = FieldFill.INSERT)
     @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime createTime;
@@ -40,7 +40,7 @@ public class BaseEntity<T> implements Serializable {
     /**
      * 更新时间
      */
-    @ApiModelProperty(value = "更新时间")
+    @Schema(description = "更新时间")
     @TableField(value = "update_time", fill = FieldFill.INSERT_UPDATE)
     @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime updateTime;
@@ -48,7 +48,7 @@ public class BaseEntity<T> implements Serializable {
     /**
      * 删除标识
      */
-    @ApiModelProperty(value = "删除标识")
+    @Schema(description = "删除标识")
     @TableField(value = "is_deleted", select = false)
     @Getter()
     private boolean deleted;
