@@ -5,7 +5,7 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
 import java.io.Serializable;
@@ -22,26 +22,26 @@ public class Music implements Serializable {
     /**
      * 业务主表主键ID
      */
-    @ApiModelProperty(value = "业务主表主键ID")
+    @Schema(description = "业务主表主键ID")
     @JsonSerialize(using = ToStringSerializer.class)
     @TableId(value = "id")
     private Long id;
     /**
      * 博客名称
      */
-    @ApiModelProperty(value = "音乐名称")
+    @Schema(description = "音乐名称")
     @TableField(value = "music_name")
     private String musicName;
     /**
      * 博客名称
      */
-    @ApiModelProperty(value = "博客名称")
+    @Schema(description = "博客名称")
     @TableField(value = "music_code")
     private String musicCode;
     /**
      * 博客名称
      */
-    @ApiModelProperty(value = "博客名称")
+    @Schema(description = "博客名称")
     @TableField(value = "music_path")
     private String musicPath;
 }
