@@ -5,7 +5,7 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
@@ -25,7 +25,7 @@ public class Picture extends BaseEntity<Picture> implements Serializable {
     /**
      * 业务主表主键ID
      */
-    @ApiModelProperty(value = "业务主表主键ID")
+    @Schema(description = "业务主表主键ID")
     @JsonSerialize(using = ToStringSerializer.class)
     @TableId(value = "id")
     private Long id;
@@ -33,28 +33,28 @@ public class Picture extends BaseEntity<Picture> implements Serializable {
     /**
      * 图片名称
      */
-    @ApiModelProperty(value = "图片名称")
+    @Schema(description = "图片名称")
     @TableField(value = "picture_name")
     private String pictureName;
 
     /**
      * 路径
      */
-    @ApiModelProperty(value = "图片路径")
+    @Schema(description = "图片路径")
     @TableField(value = "picture_address")
     private String pictureAddress;
 
     /**
      * 内容
      */
-    @ApiModelProperty(value = "图片内容")
+    @Schema(description = "图片内容")
     @TableField(value = "picture_description")
     private String pictureDescription;
 
     /**
      * 层次
      */
-    @ApiModelProperty(value = "层次")
+    @Schema(description = "层次")
     @TableField(value = "picture_level")
     private String pictureLevel;
 }
