@@ -1,7 +1,6 @@
 package com.slcp.devops.entity;
 
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
 import java.io.Serializable;
@@ -12,30 +11,30 @@ import java.io.Serializable;
  * @author devops
  */
 @Data
-@ApiModel(description = "搜索条件")
+@Schema(description = "搜索条件")
 public class Search implements Serializable {
 
     /**
      * 当前页
      */
-    @ApiModelProperty(value = "当前页")
+    @Schema(description = "当前页")
     private Integer pageNum = 1;
 
     /**
      * 每页的数量
      */
-    @ApiModelProperty(value = "每页的数量")
+    @Schema(description = "每页的数量")
     private Integer pageSize = 10;
 
     /**
      * 升序 字段数组
      */
-    @ApiModelProperty(hidden = true)
+    @Schema(hidden = true)
     private String ascs;
 
     /**
      * 降序 字段数组
      */
-    @ApiModelProperty(hidden = true)
+    @Schema(hidden = true)
     private String descs;
 }
