@@ -5,7 +5,7 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
 import java.io.Serializable;
@@ -23,34 +23,34 @@ public class SysRights implements Serializable {
     /**
      * 业务主表主键ID
      */
-    @ApiModelProperty(value = "业务主表主键ID")
+    @Schema(description = "业务主表主键ID")
     @JsonSerialize(using = ToStringSerializer.class)
     @TableId(value = "id")
     private Long id;
     /**
      * 角色id
      */
-    @ApiModelProperty(value = "角色id")
+    @Schema(description = "角色id")
     @JsonSerialize(using = ToStringSerializer.class)
     @TableField(value = "rid")
     private Long rid;
     /**
      * 权限名称
      */
-    @ApiModelProperty(value = "权限名称")
+    @Schema(description = "权限名称")
     @TableField(value = "auth_name")
     private String authName;
     /**
      * 级别
      */
-    @ApiModelProperty(value = "级别")
+    @Schema(description = "级别")
     @JsonSerialize(using = ToStringSerializer.class)
     @TableField(value = "level")
     private Long level;
     /**
      * 路径
      */
-    @ApiModelProperty(value = "路径")
+    @Schema(description = "路径")
     @TableField(value = "path")
     private String path;
 
