@@ -5,9 +5,10 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
 
@@ -18,6 +19,7 @@ import java.io.Serializable;
  * @since 2022-07-09 23:53:22
  */
 @Data
+@NoArgsConstructor
 @AllArgsConstructor
 @TableName(value = "tag_blog_rel")
 public class TagBlogRel implements Serializable {
@@ -25,7 +27,7 @@ public class TagBlogRel implements Serializable {
     /**
      * 业务主表主键ID
      */
-    @ApiModelProperty(value = "业务主表主键ID")
+    @Schema(description = "业务主表主键ID")
     @JsonSerialize(using = ToStringSerializer.class)
     @TableId(value = "id")
     private Long id;
@@ -33,14 +35,14 @@ public class TagBlogRel implements Serializable {
     /**
      * 标签id
      */
-    @ApiModelProperty(value = "标签id")
+    @Schema(description = "标签id")
     @TableField(value = "tid")
     private Long tagId;
 
     /**
      * 博客id
      */
-    @ApiModelProperty(value = "博客id")
+    @Schema(description = "博客id")
     @TableField(value = "bid")
     private Long blogId;
 
