@@ -1,13 +1,13 @@
 package com.slcp.devops.service.impl;
 
 import com.slcp.devops.config.DoQueryCache;
-import com.slcp.devops.mapper.CommentMapper;
 import com.slcp.devops.dto.Comment;
+import com.slcp.devops.mapper.CommentMapper;
 import com.slcp.devops.mapper.IBlogMapper;
 import com.slcp.devops.service.CommentService;
+import jakarta.annotation.Resource;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.Resource;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
