@@ -1,3 +1,4 @@
+package com.slcp.devops.utils;/*
 package com.slcp.devops.utils;
 
 import com.slcp.devops.entity.openai.OpenAi;
@@ -7,10 +8,15 @@ import com.theokanning.openai.completion.CompletionRequest;
 import org.springframework.util.StringUtils;
 import java.util.*;
 
+*/
 /**
  * @author slcp
  * @date 2023-01-10 22:30:21
- */
+ * @deprecated 已废弃，请使用新的 {@link com.slcp.devops.service.AiChatService} 代替
+ * 该类仅保留用于向后兼容，将在未来版本中移除
+ *//*
+
+@Deprecated
 public class OpenAiUtils {
 
     public static List<String> OPENAPI_TOKEN = Collections.singletonList("");
@@ -70,13 +76,15 @@ public class OpenAiUtils {
         PARMS.put("OpenAi49", new OpenAi("OpenAi49", "面试", "生成面试题", "text-davinci-003", "创建10道%s相关的面试题（中文）：\n", 0.5, 1.0, 10.0, 0.0, 0.0, ""));
     }
 
-    /**
+    */
+/**
      * 获取ai
      *
      * @param openAi
      * @param prompt
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getAiResult(OpenAi openAi, String prompt) {
         if (TIMEOUT == null || TIMEOUT < 1000) {
             TIMEOUT = 3000;
@@ -97,67 +105,79 @@ public class OpenAiUtils {
         return service.createCompletion(completionRequest).getChoices();
     }
 
-    /**
+    */
+/**
      * 问答
      *
      * @param question
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getQuestionAnswer(String question) {
         OpenAi openAi = PARMS.get("OpenAi01");
         return getAiResult(openAi, String.format(openAi.getPrompt(), question));
     }
 
-    /**
+    */
+/**
      * 语法纠错
      *
      * @param text
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getGrammarCorrection(String text) {
         OpenAi openAi = PARMS.get("OpenAi02");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 将一段话，概况中心
      *
      * @param text
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getSummarize(String text) {
         OpenAi openAi = PARMS.get("OpenAi03");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 一句话生成OpenAi的代码
      *
      * @param text
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getOpenAiApi(String text) {
         OpenAi openAi = PARMS.get("OpenAi04");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 一句话生成程序的命令，目前支持操作系统指令比较多
      *
      * @param text
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getTextToCommand(String text) {
         OpenAi openAi = PARMS.get("OpenAi05");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 把一种语法翻译成其它几种语言
      *
      * @param text
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getTranslatesLanguages(String text, String translatesLanguages) {
         if (StringUtils.isEmpty(translatesLanguages)) {
             translatesLanguages = "  1. French, 2. Spanish and 3. English";
@@ -166,19 +186,22 @@ public class OpenAiUtils {
         return getAiResult(openAi, String.format(openAi.getPrompt(), translatesLanguages, text));
     }
 
-    /**
+    */
+/**
      * 一句话生成Stripe国际支付API
      *
      * @param text
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getStripeApi(String text) {
         OpenAi openAi = PARMS.get("OpenAi07");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
 
-    /**
+    */
+/**
      * 依据上下文中的表信息，生成SQL语句
      *
      * @param databaseType 数据库类型
@@ -186,7 +209,8 @@ public class OpenAiUtils {
      * @param text         SQL描述
      * @param sqlType      sql类型，比如SELECT
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getStripeApi(String databaseType, List<String> tables, String text, String sqlType) {
         OpenAi openAi = PARMS.get("OpenAi08");
         StringJoiner joiner = new StringJoiner("\n");
@@ -196,82 +220,96 @@ public class OpenAiUtils {
         return getAiResult(openAi, String.format(openAi.getPrompt(), databaseType, joiner.toString(), text, sqlType));
     }
 
-    /**
+    */
+/**
      * 对于非结构化的数据抽取其中的特征生成结构化的表格
      *
      * @param text 非结构化的数据
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getUnstructuredData(String text) {
         OpenAi openAi = PARMS.get("OpenAi09");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 把一段信息继续分类
      *
      * @param text 要分类的文本
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getTextCategory(String text) {
         OpenAi openAi = PARMS.get("OpenAi10");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 把一段信息继续分类
      *
      * @param codeType 代码类型，比如Python
      * @param code     要解释的代码
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getCodeExplain(String codeType, String code) {
         OpenAi openAi = PARMS.get("OpenAi11");
         return getAiResult(openAi, String.format(openAi.getPrompt(), codeType, code));
     }
 
-    /**
+    */
+/**
      * 将文本编码成表情服务
      *
      * @param text 文本
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getTextEmoji(String text) {
         OpenAi openAi = PARMS.get("OpenAi12");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 求一段代码的时间复杂度
      *
      * @param code 代码
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getTimeComplexity(String code) {
         OpenAi openAi = PARMS.get("OpenAi13");
         return getAiResult(openAi, String.format(openAi.getPrompt(), code));
     }
 
 
-    /**
+    */
+/**
      * 把一种语言的代码翻译成另外一种语言的代码
      *
      * @param fromLanguage 要翻译的代码语言
      * @param toLanguage   要翻译成的代码语言
      * @param code         代码
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getTranslateProgramming(String fromLanguage, String toLanguage, String code) {
         OpenAi openAi = PARMS.get("OpenAi14");
         return getAiResult(openAi, String.format(openAi.getPrompt(), fromLanguage, toLanguage, fromLanguage, code, toLanguage));
     }
 
-    /**
+    */
+/**
      * 支持批量列表的方式检查情绪
      *
      * @param texts 文本
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getBatchTweetClassifier(List<String> texts) {
         OpenAi openAi = PARMS.get("OpenAi15");
         StringJoiner stringJoiner = new StringJoiner("\n");
@@ -281,320 +319,376 @@ public class OpenAiUtils {
         return getAiResult(openAi, String.format(openAi.getPrompt(), stringJoiner.toString()));
     }
 
-    /**
+    */
+/**
      * 对一段代码进行解释
      *
      * @param code 文本
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getExplainCOde(String code) {
         OpenAi openAi = PARMS.get("OpenAi16");
         return getAiResult(openAi, String.format(openAi.getPrompt(), code));
     }
 
-    /**
+    */
+/**
      * 提取一段文本中的关键字
      *
      * @param text 文本
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getTextKeywords(String text) {
         OpenAi openAi = PARMS.get("OpenAi17");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 事实回答答题
      *
      * @param text 文本
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getFactualAnswering(String text) {
         OpenAi openAi = PARMS.get("OpenAi18");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 给一个产品设计一个广告
      *
      * @param text 文本
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getAd(String text) {
         OpenAi openAi = PARMS.get("OpenAi19");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 依据产品描述和种子词语，给一个产品取一个好听的名字
      *
      * @param productDescription 产品描述
      * @param seedWords          种子词语
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getProductName(String productDescription, String seedWords) {
         OpenAi openAi = PARMS.get("OpenAi20");
         return getAiResult(openAi, String.format(openAi.getPrompt(), productDescription, seedWords));
     }
 
-    /**
+    */
+/**
      * 把一个长句子简化成一个短句子
      *
      * @param text 长句子
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getProductName(String text) {
         OpenAi openAi = PARMS.get("OpenAi21");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 自动修改代码中的bug
      *
      * @param codeType 语言类型
      * @param code     代码
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getBugFixer(String codeType, String code) {
         OpenAi openAi = PARMS.get("OpenAi22");
         return getAiResult(openAi, String.format(openAi.getPrompt(), codeType, code, codeType));
     }
 
-    /**
+    */
+/**
      * 自动为一个表格生成数据
      *
      * @param rows    生成的行数
      * @param headers 数据表头，格式如：姓名| 年龄|性别|生日
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getFillData(int rows, String headers) {
         OpenAi openAi = PARMS.get("OpenAi23");
         return getAiResult(openAi, String.format(openAi.getPrompt(), rows, headers));
     }
 
-    /**
+    */
+/**
      * 各种开发语言的两天机器人
      *
      * @param question             你的问题
      * @param programmingLanguages 语言 比如Java JavaScript
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getProgrammingLanguageChatbot(String question, String programmingLanguages) {
         OpenAi openAi = PARMS.get("OpenAi24");
         return getAiResult(openAi, String.format(openAi.getPrompt(), question, programmingLanguages));
     }
 
-    /**
+    */
+/**
      * 机器学习模型方面的机器人
      *
      * @param question 你的问题
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getMLChatbot(String question) {
         OpenAi openAi = PARMS.get("OpenAi25");
         return getAiResult(openAi, String.format(openAi.getPrompt(), question));
     }
 
-    /**
+    */
+/**
      * 可以列出各方面的分类列表，比如歌单
      *
      * @param text 清单描述
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getListMaker(String text) {
         OpenAi openAi = PARMS.get("OpenAi26");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 对一段文字进行情绪分析
      *
      * @param text
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getTweetClassifier(String text) {
         OpenAi openAi = PARMS.get("OpenAi27");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 抽取文本中的航空代码信息
      *
      * @param text
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getAirportCodeExtractor(String text) {
         OpenAi openAi = PARMS.get("OpenAi28");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 无上下文，语句描述生成SQL
      *
      * @param text
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getSQL(String text) {
         OpenAi openAi = PARMS.get("OpenAi29");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 从文本中抽取联系方式
      *
      * @param extractContent 抽取内容描述
      * @param text
      * @return 从下面文本中抽取邮箱和电话:\n教育行业A股IPO第一股（股票代码 003032）\n全国咨询/投诉热线：400-618-4000    举报邮箱:mc@itcast.cn
-     */
+     *//*
+
     public static List<CompletionChoice> getExtractContactInformation(String extractContent, String text) {
         OpenAi openAi = PARMS.get("OpenAi30");
         return getAiResult(openAi, String.format(openAi.getPrompt(), extractContent, text));
     }
 
-    /**
+    */
+/**
      * 把一种语言转成另外一种语言代码
      *
      * @param fromCodeType 当前代码类型
      * @param toCodeType   转换的代码类型
      * @param code
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getTransformationCode(String fromCodeType, String toCodeType, String code) {
         OpenAi openAi = PARMS.get("OpenAi31");
         return getAiResult(openAi, String.format(openAi.getPrompt(), fromCodeType, toCodeType, fromCodeType, code, toCodeType));
     }
 
-    /**
+    */
+/**
      * 模仿好友聊天
      *
      * @param question
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getFriendChat(String question) {
         OpenAi openAi = PARMS.get("OpenAi32");
         return getAiResult(openAi, String.format(openAi.getPrompt(), question));
     }
 
-    /**
+    */
+/**
      * 依据描述生成对应颜色
      *
      * @param text
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getMoodToColor(String text) {
         OpenAi openAi = PARMS.get("OpenAi33");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 自动为程序生成文档
      *
      * @param codeType 语言
      * @param code
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getCodeDocument(String codeType, String code) {
         OpenAi openAi = PARMS.get("OpenAi34");
         return getAiResult(openAi, String.format(openAi.getPrompt(), codeType, code));
     }
 
-    /**
+    */
+/**
      * 依据短语生成相关文短
      *
      * @param text 短语
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getCreateAnalogies(String text) {
         OpenAi openAi = PARMS.get("OpenAi35");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 把多行代码简单的压缩成一行
      *
      * @param codeType 语言
      * @param code
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getCodeLine(String codeType, String code) {
         OpenAi openAi = PARMS.get("OpenAi36");
         return getAiResult(openAi, String.format(openAi.getPrompt(), codeType, code, codeType));
     }
 
-    /**
+    */
+/**
      * 依据一个主题创建一个故事
      *
      * @param topic 创作主题
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getStory(String topic) {
         OpenAi openAi = PARMS.get("OpenAi37");
         return getAiResult(openAi, String.format(openAi.getPrompt(), topic));
     }
 
-    /**
+    */
+/**
      * 第一人称转第3人称
      *
      * @param text
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getStoryCreator(String text) {
         OpenAi openAi = PARMS.get("OpenAi38");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 依据笔记生成摘要说明
      *
      * @param scene 生成的摘要场景
      * @param note  记录的笔记
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getNotesToSummary(String scene, String note) {
         OpenAi openAi = PARMS.get("OpenAi39");
         return getAiResult(openAi, String.format(openAi.getPrompt(), note));
     }
 
-    /**
+    */
+/**
      * 给定一个主题，让其生成一些主题相关的想法
      *
      * @param topic 头脑风暴关键词
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getIdeaGenerator(String topic) {
         OpenAi openAi = PARMS.get("OpenAi40");
         return getAiResult(openAi, String.format(openAi.getPrompt(), topic));
     }
 
-    /**
+    */
+/**
      * 按照ESRB进行文本分类
      *
      * @param text 文本
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getESRBRating(String text) {
         OpenAi openAi = PARMS.get("OpenAi41");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 按照提示为相关内容生成提纲
      *
      * @param text 场景，比如 数据库软件生成大学毕业论文
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getEssayOutline(String text) {
         OpenAi openAi = PARMS.get("OpenAi42");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 依据美食名称和材料生成美食的制作步骤
      *
      * @param name        美食名称
      * @param ingredients 美食食材
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getRecipeCreator(String name, List<String> ingredients) {
         OpenAi openAi = PARMS.get("OpenAi43");
         StringJoiner joiner = new StringJoiner("\n");
@@ -604,71 +698,83 @@ public class OpenAiUtils {
         return getAiResult(openAi, String.format(openAi.getPrompt(), name, joiner.toString()));
     }
 
-    /**
+    */
+/**
      * 与AI机器进行聊天
      *
      * @param question
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getAiChatbot(String question) {
         OpenAi openAi = PARMS.get("OpenAi44");
         return getAiResult(openAi, String.format(openAi.getPrompt(), question));
     }
 
-    /**
+    */
+/**
      * 与讽刺机器进行聊天，聊天的机器人是一种消极情绪
      *
      * @param question
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getMarvChatbot(String question) {
         OpenAi openAi = PARMS.get("OpenAi45");
         return getAiResult(openAi, String.format(openAi.getPrompt(), question));
     }
 
-    /**
+    */
+/**
      * 把一段文本分解成几步来完成
      *
      * @param text
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getTurnDirection(String text) {
         OpenAi openAi = PARMS.get("OpenAi46");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 依据文本内容自动生成点评
      *
      * @param text
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getReviewCreator(String text) {
         OpenAi openAi = PARMS.get("OpenAi47");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 可以为学习知识自动解答
      *
      * @param text
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getStudyNote(String text) {
         OpenAi openAi = PARMS.get("OpenAi48");
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-    /**
+    */
+/**
      * 生成面试题
      *
      * @param text
      * @return
-     */
+     *//*
+
     public static List<CompletionChoice> getInterviewQuestion(String text) {
         OpenAi openAi = PARMS.get("OpenAi49");
         System.out.println(String.format(openAi.getPrompt(), text));
         return getAiResult(openAi, String.format(openAi.getPrompt(), text));
     }
 
-}
+}*/
