@@ -1,9 +1,9 @@
 package com.slcp.devops.websocket;
 
+import jakarta.websocket.*;
+import jakarta.websocket.server.ServerEndpoint;
 import org.springframework.stereotype.Component;
 
-import javax.websocket.*;
-import javax.websocket.server.ServerEndpoint;
 import java.io.IOException;
 import java.util.concurrent.CopyOnWriteArrayList;
 
